requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
bcrypt==4.1.2
python-multipart==0.0.6
jinja2==3.1.2